v8

+++ app/Http/Controllers/CategoriaController.php
@@ -6,7 +6,6 @@ use App\Http\Requests\Categoria\StoreRequest;
 use App\Http\Requests\Categoria\UpdateRequest;
 use App\Models\Categoria;
 use Illuminate\Http\Request;
-use Laracasts\Flash\Flash;
 
 class CategoriaController extends Controller
 {
@@ -29,8 +28,7 @@ class CategoriaController extends Controller
         $categoria->nombre = $request->nombre;
         $categoria->descripcion = $request->descripcion;
         $categoria->save();
-        flash::success('La categoria fue registrada con éxito');
-        return redirect ('/categorias');
+        return redirect()->route('categorias.index');
     }
     public function show(Categoria $categoria)
     {
@@ -46,14 +44,18 @@ class CategoriaController extends Controller
         $categoria->nombre = $request->nombre;
         $categoria->descripcion = $request->descripcion;
         $categoria->update();
-        flash::warning('La categoria fue actualizada con éxito');
-        return redirect ('/categorias');
+        return redirect()->route('categorias.index');
     }
     public function destroy(Categoria $categoria)
     {
         
-        $categoria->delete();
-        flash::error('La categoria fue eliminada con éxito');
-        return redirect('/categorias');
+        $fk =  $categoria->fk_categoriaProducto($categoria->id);
+        if($fk == 1)
+        {
+            return redirect()->route('categorias.index')->with('eliminar','error');
+        }else{
+            $categoria->delete();
+            return redirect()->route('categorias.index')->with('eliminar','ok');
+        }
     }
 }
